agents/test-client/pyproject.toml
[project]
name = "test-client"
version = "0.1.0"
description = "HelloWorld client agent that retrieves messages from HelloWorld server agent"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "httpx>=0.28.1",
    "httpx-auth>=0.23.1",
    "hvac>=2.4.0",
]
